// src/app/Application/Modules/WPSPayroll/model/responseListProcessedFilesPayrollWPS.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { ErrorServiceResponse } from './errorServiceResponse'
import { PagedResultsWPSSalaryFileDSO } from './pagedResultsWPSSalaryFileDSO'
import { PayrollParametersDSO } from './payrollParametersDSO'
import { ResponseGenerateChallenge } from 'app/Application/Model/responsegeneratechallenge.type'
import { Account } from 'app/Application/Model/account'

export interface ResponseListProcessedFilesPayrollWPS {
  accountListLocal?: Array<Account>
  accountSelected?: string
  errorCode?: string
  errorDescription?: string
  errorResponse?: ErrorServiceResponse
  generateChallengeAndOTP?: ResponseGenerateChallenge
  listFile?: PagedResultsWPSSalaryFileDSO
  payrollCompanyDetails?: PayrollParametersDSO
}